- Clean up some more the select target page. - Remove `AvailableTargetVariables` component. - Include `VariableFacets` component directly in the select target page.
